Fix `expo-contacts/next` throwing `Cannot find native module 'ExpoContactsNext'`
